requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
tenacity==8.2.3
python-dotenv==1.0.1
lxml==5.1.0
google-search-results==2.4.2
playwright>=1.44.0
crawl4ai>=0.4.0
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.27.0
peft>=0.10.0
openai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
